mask.go: Reorganize
